[#248]refactor: context 및 hook 분리
